Extract `JSONSharedTestCases`
Both `mysql2/json_test.rb` and `postgresql/json_test.rb` have same test
cases.
